+++ src/ReshareRequest.js
@@ -116,6 +116,7 @@ function translateCOtoRR(co) {
 
   // These are non-standard fields in OpenURL 1.0
   rr.serviceType = _.get(a, 'svc.id'); // No example of this in Z39.88
+  rr.serviceLevel = _.get(m, 'svc.level');
   rr.isRequester = true;
   const copyrightType = _.get(m, 'rft.copyrightType');
   if (copyrightType) rr.copyrightType = { value: copyrightType };